src/proposal_generator/components/market_analyzer.py
"""Market Analyzer module."""
from typing import Dict, Any, List
import logging

logger = logging.getLogger(__name__)

class MarketAnalyzer:
    """Class for analyzing market data and trends."""

    async def analyze_market(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Analyze market data and generate insights.
        
        Args:
            data: Dictionary containing:
                - industry: Industry type
                - target_audience: List of target audience segments
                - competitors: List of competitor URLs
                - key_services: List of key services/products
                
        Returns:
            Dict containing market analysis results
        """
        try:
            # Extract data with proper type checking
            industry = str(data.get('industry', ''))
            target_audience = data.get('target_audience', [])
            competitors = data.get('competitors', [])
            key_services = data.get('key_services', [])

            if not isinstance(target_audience, list):
                target_audience = [str(target_audience)]
            if not isinstance(competitors, list):
                competitors = [str(competitors)]
            if not isinstance(key_services, list):
                key_services = [str(key_services)]

            # Generate market trends
            trends = self._generate_trends(industry, key_services)

            # Analyze competitors
            competitor_analysis = []
            for url in competitors:
                if url:  # Only analyze non-empty URLs
                    competitor_analysis.append(self._analyze_competitor(url))

            # Generate recommendations
            recommendations = self._generate_recommendations(
                industry=industry,
                trends=trends,
                competitor_analysis=competitor_analysis
            )

            return {
                'industry': industry,
                'trends': trends,
                'competitors': competitor_analysis,
                'recommendations': recommendations
            }

        except Exception as e:
            logger.error(f"Error in market analysis: {str(e)}")
            return {
                'industry': '',
                'trends': [],
                'competitors': [],
                'recommendations': [
                    {
                        'priority': 'high',
                        'description': f'Error during market analysis: {str(e)}'
                    }
                ]
            }

    def _generate_trends(self, industry: str, key_services: List[str]) -> List[Dict[str, Any]]:
        """Generate industry trends."""
        # Placeholder implementation
        return [
            {
                'name': 'Digital Transformation',
                'description': 'Increasing adoption of digital solutions',
                'adoption_rate': 0.85,
                'impact': 'High'
            },
            {
                'name': 'Sustainability',
                'description': 'Growing focus on environmental impact',
                'adoption_rate': 0.75,
                'impact': 'Medium'
            }
        ]

    def _analyze_competitor(self, url: str) -> Dict[str, Any]:
        """Analyze a competitor's website."""
        # Placeholder implementation
        return {
            'name': url.split('//')[1].split('/')[0],
            'market_share': 0.15,
            'strengths': ['brand recognition', 'product quality']
        }

    def _generate_recommendations(self, industry: str, trends: List[Dict[str, Any]], 
                                competitor_analysis: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Generate market-based recommendations."""
        recommendations = []

        # Add trend-based recommendations
        for trend in trends:
            recommendations.append({
                'priority': 'high' if trend.get('adoption_rate', 0) > 0.8 else 'medium',
                'description': f"Capitalize on {trend['name']} trend: {trend['description']}"
            })

        # Add competitor-based recommendations
        if competitor_analysis:
            recommendations.append({
                'priority': 'high',
                'description': 'Differentiate from competitors through unique value proposition'
            })

        return recommendations